requirements.txt
absl-py==2.0.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.0
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
backoff==2.2.1
blobfile==2.1.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.5.2
colorama==0.4.6
compel==2.0.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.7
dctorch==0.1.2
Deprecated==1.2.14
diffusers==0.24.0
docarray==0.39.1
docker==6.1.3
docker-pycreds==0.4.0
ecdsa==0.18.0
einops==0.3.0
fastapi==0.104.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-health-checking==1.59.3
grpcio-reflection==1.59.3
guided-diffusion-sdk==1.0.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.0
importlib-metadata==6.8.0
jcloud==0.3
jina==3.23.0
jina-hubble-sdk==0.39.0
Jinja2==3.1.2
joblib==1.3.2
jsonmerge==1.9.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
k-diffusion==0.1.1
keras==2.15.0
kiwisolver==1.4.5
kornia==0.6.0
lapx==0.5.5
lazy_loader==0.3
libclang==16.0.6
lpips==0.1.4
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.0.2
openai-clip==1.0.1
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-grpc==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
opentelemetry-util-http==0.40b0
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
Pillow==10.1.0
prometheus-client==0.19.0
protobuf==4.23.4
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.19.0
pydantic==1.10.13
pyDeprecate==0.3.1
pygame==2.5.2
Pygments==2.17.1
pyparsing==3.1.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-version==0.0.2
pytorch-lightning==1.4.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right-sdk==1.0.2
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
sentencepiece==0.1.99
sentry-sdk==1.36.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stable-diffusion==0.0.0
stable-diffusion-sdkit==2.1.5
starlette==0.27.0
strongsort==0.1.2
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
test-tube==0.7.5
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.11
tokenizers==0.13.3
torch==2.1.2
torchdiffeq==0.2.3
torchgeometry==0.1.2
torchmetrics==0.6.0
torchsde==0.2.6
torchvision==0.16.2
tqdm==4.66.1
trampoline==0.1.2
transformers==4.33.2
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.234
urllib3==1.26.18
uvicorn==0.23.1
wandb==0.16.0
watchfiles==0.21.0
wcwidth==0.2.11
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
xgboost==2.0.3
yarl==1.9.3
zipp==3.17.0
